+++ castle/cms/browser/controlpanel/archival.py
@@ -43,6 +43,9 @@ class Review(BaseView):
         return json.dumps(self.items())
 
     def __call__(self):
+        # utility function to add resource to rendered page
+        add_resource_on_request(self.request, 'castle-components-review-archives')
+
         extend = self.request.form.get('extend')
         if extend:
             obj = uuidToObject(extend)
@@ -57,6 +60,7 @@ class AWSApi(object):
         self.site = site
         self.request = request
         bucket_name = api.portal.get_registry_record('castle.aws_s3_bucket_name')
+        self.configured_bucket_name = bucket_name
         self.s3, self.bucket = aws.get_bucket(bucket_name)
         self.archive_storage = archival.Storage(site)
 
@@ -172,17 +176,48 @@ class AWSApi(object):
         result = []
         base_path = archival.CONTENT_KEY_PREFIX + self.request.form.get('path', '')
         base_path = base_path.replace('//', '/').rstrip('/') + '/'
-        for key in self.bucket.objects.filter(prefix=base_path):
-            path = key.name[len(archival.CONTENT_KEY_PREFIX):]
+
+        if self.bucket is None:
+            logger.error('no bucket object (configured bucket name: {})'.format(self.configured_bucket_name))
+            return []
+
+        try:
+            page = int(self.request.form.get('page', 1))
+        except Exception:
+            page = 1
+        if page <= 0:
+            page = 1
+        per_page = 1000
+        page_start = (page - 1) * per_page
+        page_end = page * per_page
+        s3_per_page = 1000
+
+        # this is so we don't kill the client with a large amount of objects being sent that way
+        # note: s3 api doesn't do paging like the client would expect to render. the pages defined
+        # for the api are basically equivalent to 1000 item max chunks retreived from the bucket.
+        #
+        # an improvement on this would probably be caching the list with every publish to the
+        # archive and having somewhere to manually update the cache. TODO.
+
+        object_itr = self.bucket.objects.page_size(count=s3_per_page).filter(Prefix=base_path)
+        i = 0
+        for summary in object_itr:
+            i += 1
+            if i < page_start:
+                continue
+            key = summary.key
+            path = key[len(archival.CONTENT_KEY_PREFIX):]
             result.append({
                 'path': path,
                 'id': path.rstrip('/').split('/')[-1],
                 'is_folder': False,
                 'url': '{endpoint_url}/{bucket}/{key}'.format(
-                    endpoint_url=self.s3.meta.endpoint_url,
+                    endpoint_url=self.s3.meta.client.meta.endpoint_url,
                     bucket=self.bucket.name,
-                    key=archival.CONTENT_KEY_PREFIX + path)
+                    key=key)
             })
+            if i >= page_end:
+                break
         return result
 
 
@@ -191,7 +226,6 @@ class Manage(BaseView):
     def __call__(self):
         # utility function to add resource to rendered page
         add_resource_on_request(self.request, 'castle-components-manage-archives')
-        return super(Manage, self).__call__()
 
         if self.request.form.get('api'):
             self.request.response.setHeader('Content-type', 'application/json')